The network configuration now makes more sense
